+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## x
+- FEAT: `k.once` matching with chaining (will deprecate `k.on`)
 - FEAT: Ruby 2.1 compatibility
 - FEAT: `k.when` returns `k` so when's can be defined as a chain
 - FIX: Ruby 2.4 Fixnum deprecation warning